# JDK-8299158: Improve MD5 intrinsic on AArch64